% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{buildcorrsigned}
\alias{buildcorrsigned}
\title{Create the correlation matrix (used internally)}
\usage{
buildcorrsigned(id1, id2, type)
}
\arguments{
\item{id1}{focal id}

\item{id2}{other id}

\item{type}{type of interaction}
}
\value{
a list with pairsID, correlation Matrix, and decomposed matrices
}
\description{
Create the correlation matrix (used internally)
}
