% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{create_DF_sex}
\alias{create_DF_sex}
\title{Create dataset sex}
\usage{
create_DF_sex(DF1)
}
\arguments{
\item{DF1}{The processed data for different sex}
}
\value{
a DF with females as focal
}
\description{
Creates the data use for the model with female as focal,
it reverses all interactions in which the focal is a male.
}
\examples{
data(data_diff_sex)
diff_sex_sex <- create_DF_sex(data_diff_sex)

}
