% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{create_DF_social}
\alias{create_DF_social}
\title{Create dataset social support}
\usage{
create_DF_social(DF1)
}
\arguments{
\item{DF1}{The processed data for same sex or different sex}
}
\value{
a DF with higest social support as focal
}
\description{
this function create the data use for the model with social support as focal,
it reverses all interactions in which the focal has a lower social support.
}
\examples{
data(data_same_sex)
data(data_diff_sex)
data(data_resid)
same_sex_social <- create_DF_social(data_same_sex)
diff_sex_social <- create_DF_social(data_diff_sex)
resid_social <- create_DF_social(data_resid)

}
