% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{create_DF_weight}
\alias{create_DF_weight}
\title{Create dataset body mass}
\usage{
create_DF_weight(DF1)
}
\arguments{
\item{DF1}{The processed data for same sex or different sex}
}
\value{
a DF with heaviest as focal
}
\description{
Create the data use for the model with body mass as focal.
it reverses all interaction in which the focal has a lower body mass than its
opponent.
}
\examples{
data(data_diff_sex)
data(data_same_sex)
diff_sex_weight <- create_DF_weight(data_diff_sex)
same_sex_weight <- create_DF_weight(data_same_sex)

}
