% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_diff_sex}
\alias{data_diff_sex}
\title{different sex data}
\format{A data frame with 1109 interactions and 10 variables}
\usage{
data(data_diff_sex)
}
\description{
A dataset containing the interaction of all different sex interactions
}
\keyword{datasets}
