% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_resid}
\alias{data_resid}
\title{data resid}
\format{A data frame with 153 interactiosn and 10 variables}
\usage{
data(data_resid)
}
\description{
A dataset containing the interclan interactions in which the migrants have
the highest social support
}
\keyword{datasets}
