% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_same_sex}
\alias{data_same_sex}
\title{same sex data}
\format{
A data frame with 3024 interactions and 10 variables
}
\usage{
data(data_same_sex)
}
\description{
A dataset containing the interaction of all same sex interactions
}
\keyword{datasets}
