% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diff_sex_sex}
\alias{diff_sex_sex}
\title{different sex sex}
\format{A data frame with 1109 interactions and 10 variables}
\usage{
data(diff_sex_sex)
}
\description{
A dataset containing the interaction of all different sex interaction with
the individuals with the female as focal
}
\keyword{datasets}
