% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diff_sex_social}
\alias{diff_sex_social}
\title{different sex social}
\format{A data frame with 1109 interactiosn and 10 variables}
\usage{
data(diff_sex_social)
}
\description{
A dataset containing the interaction of all different sex interaction with
the individuals with the highest social support as focal
}
\keyword{datasets}
