% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{females_relat}
\alias{females_relat}
\title{PLOT: females relat}
\usage{
data(females_relat)
}
\description{
A dataset containing the relatedness score for females at 3 different ages
// used for plot1
}
\keyword{datasets}
