% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_body_mass}
\alias{fit_body_mass}
\title{Fit body mass models}
\usage{
fit_body_mass(fit_method, model, DF1)
}
\arguments{
\item{fit_method}{"PQL" or "PQL/L" in quotes}

\item{model}{"full", "nosocial", "nosex", "null" in quotes}

\item{DF1}{dataset to use: same_sex_weight or diff_sex_weight}
}
\value{
a list (fitted model with spaMM)
}
\description{
Fits the models when the heaviest individual is focal.
}
\examples{
\dontrun{
mod_mass_full_same_PQL <- fit_body_mass(fit_method = "PQL",
                                       model = "full",
                                       DF1 = same_sex_weight)
mod_mass_full_diff_PQL <- fit_social(fit_method = "PQL",
                                    model = "full",
                                    DF1 = diff_sex_weight)
                             
mod_mass_full_same_PQLL <- fit_social(fit_method = "PQL/L",
                                     model = "full",
                                     DF1 = same_sex_weight)
mod_mass_full_diff_PQLL <- fit_social(fit_method = "PQL/L",
                                     model = "full",
                                     DF1 = diff_sex_weight)
                             
mod_mass_nosocial_same_PQL <- fit_social(fit_method = "PQL",
                                        model = "nosocial",
                                        DF1 = same_sex_weight)
mod_mass_nosocial_diff_PQL <- fit_social(fit_method = "PQL",
                                        model = "nosocial",
                                        DF1 = diff_sex_weight)
                             
mod_mass_nosocial_same_PQLL <- fit_social(fit_method = "PQL/L",
                                         model = "nosocial",
                                         DF1 = same_sex_weight)
mod_mass_nosocial_diff_PQLL <- fit_social(fit_method = "PQL/L",
                                         model = "nosocial",
                                         DF1 = diff_sex_weight)
                             
mod_mass_nosex_same_PQL <- fit_social(fit_method = "PQL",
                                     model = "nosex",
                                     DF1 = same_sex_weight)
mod_mass_nosex_diff_PQL <- fit_social(fit_method = "PQL",
                                     model = "nosex",
                                     DF1 = diff_sex_weight)
                             
mod_mass_nosex_same_PQLL <- fit_social(fit_method = "PQL/L",
                                      model = "nosex",
                                      DF1 = same_sex_weight)
mod_mass_nosex_diff_PQLL <- fit_social(fit_method = "PQL/L",
                                      model = "nosex",
                                      DF1 = diff_sex_weight)
                             
mod_mass_null_same_PQL <- fit_social(fit_method = "PQL",
                                    model = "null",
                                    DF1 = same_sex_weight)
mod_mass_null_diff_PQL <- fit_social(fit_method = "PQL",
                                    model = "null",
                                    DF1 = diff_sex_weight)
                             
mod_mass_null_same_PQLL <- fit_social(fit_method = "PQL/L",
                                     model = "null",
                                     DF1 = same_sex_weight)
mod_mass_null_diff_PQLL <- fit_social(fit_method = "PQL/L",
                                     model = "null",
                                     DF1 = diff_sex_weight)
}
}
