% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_resid}
\alias{fit_resid}
\title{fit the residency model}
\usage{
fit_resid(fit_method, DF1)
}
\arguments{
\item{fit_method, }{"PQL" or "PQL/L"}

\item{DF1, }{dataset to use: resid_social}
}
\value{
a list (fitted model with spaMM)
}
\description{
this function fits the models when the migrants have higher social support than
the native in interclan context
}
\examples{
\dontrun{
resid_mod <- fit_resid(fit_method = "PQL", DF1 = resid_social)
}
}
