% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_sex}
\alias{fit_sex}
\title{Fit the models when the female is focal}
\usage{
fit_sex(model, fit_method, DF1)
}
\arguments{
\item{model}{"full", "nosocial", "nomass", "null" in quotes}

\item{fit_method}{"PQL" or "PQL/L" in quotes}

\item{DF1}{dataset to use: diff_sex_sex}
}
\value{
a list (fitted model with spaMM)
}
\description{
this function fits the models when the female is focal.
}
\examples{
\dontrun{
data(diff_sex_sex)
mod_sex_full_diff_PQL <- fit_sex(fit_method = "PQL",
                                model = "full",
                                DF1 = diff_sex_sex)
mod_sex_full_diff_PQLL <- fit_sex(fit_method = "PQL/L",
                                 model = "full",
                                 DF1 = diff_sex_sex)

mod_sex_nosocial_diff_PQL <-fit_sex(fit_method = "PQL",
                                   model = "nosocial",
                                   DF1 = diff_sex_sex)
mod_sex_nosocial_diff_PQLL <- fit_sex(fit_method = "PQL/L",
                                     model = "nosocial",
                                     DF1 = diff_sex_sex)

mod_sex_nomass_diff_PQL <- fit_sex(fit_method = "PQL",
                                  model = "nomass",
                                  DF1 = diff_sex_sex)
mod_sex_nomass_diff_PQLL <- fit_sex(fit_method = "PQL/L",
                                   model = "nomass",
                                   DF1 = diff_sex_sex)

mod_sex_null_diff_PQL <- fit_sex(fit_method = "PQL",
                                model = "null",
                                DF1 = diff_sex_sex)
mod_sex_null_diff_PQLL <- fit_sex(fit_method = "PQL/L",
                                 model = "null",
                                 DF1 = diff_sex_sex)
}

}
