% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{flip_DF_sex}
\alias{flip_DF_sex}
\title{Flip sex (used internally)}
\usage{
flip_DF_sex(DF1)
}
\arguments{
\item{DF1}{The processed data for different sex}
}
\value{
a DF with females as focal
}
\description{
Reverses all interactions in which the focal is a male
}
