% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{flip_DF_social}
\alias{flip_DF_social}
\title{Flip social support (to be used internally)}
\usage{
flip_DF_social(DF1)
}
\arguments{
\item{DF1}{data frame}
}
\value{
a df with highest social support as focal
}
\description{
Reverses all interactions in which the focal has a lower social support.
}
