% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{flip_DF_weight}
\alias{flip_DF_weight}
\title{flip body mass (used internally)}
\usage{
flip_DF_weight(DF1)
}
\arguments{
\item{DF1}{The processed data for same sex or different sex}
}
\value{
a DF with heaviest as focal
}
\description{
Reverse all interaction in which the focal has a lower body mass than its
opponent
}
