% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_AIC}
\alias{get_AIC}
\title{AIC}
\usage{
get_AIC(model, name)
}
\arguments{
\item{model}{model to be used}

\item{name}{name of the model}
}
\value{
a tibble: be aware of the automatic rounding
}
\description{
Wrapper around the spaMM::AIC method.
}
\note{
Returns a tibble so be aware of the auto-
matic rounding.
}
\examples{
data(mod_social_null_diff_PQL)
get_AIC(mod_social_null_diff_PQL, "social_null_diff_PQL")

}
