% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_and_clean_prediction}
\alias{get_and_clean_prediction}
\title{get_and_clean_prediction}
\usage{
get_and_clean_prediction(mod)
}
\arguments{
\item{mod}{social model different sex}
}
\description{
get the prediction and frame them for the plot1
}
