% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_logLik}
\alias{get_logLik}
\title{log likelihood}
\usage{
get_logLik(model, name)
}
\arguments{
\item{model}{model to be used}

\item{name}{name of the model}
}
\description{
Wrapper arround logLik.
}
\note{
Returns a tibble, so be aware of the automatic rounding.
}
\examples{
data(mod_sex_null_diff_PQL)
get_logLik(mod_sex_null_diff_PQL, "sex_null_PQL")

}
