% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_TJUR}
\alias{get_TJUR}
\title{Tjur D}
\usage{
get_TJUR(model, name)
}
\arguments{
\item{model}{model to be used}

\item{name}{name of the model}
}
\description{
Compute the Tjur's coefficient of discrimination D.
}
\examples{
data(mod_sex_null_diff_PQL)
get_TJUR(mod_sex_null_diff_PQL, "sex_null_PQL")
}
