% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migrants_relat}
\alias{migrants_relat}
\title{PLOT: migrants relat}
\usage{
data(natives_relat)
}
\description{
A dataset containing the relatedness score for migrants males at 3
different ages // used for plot1
}
\keyword{datasets}
