% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_mass_null_diff_PQL}
\alias{mod_mass_null_diff_PQL}
\title{MODEL: mod_mass_null_diff_PQL}
\usage{
data(mod_mass_null_diff_PQL)
}
\description{
Model when the individul with the highest body mass is focal
in intersex interactions.
the model is null (intercept and type of interaction) and is fitted with PQL
}
\keyword{datasets}
