% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_sex_null_diff_PQL}
\alias{mod_sex_null_diff_PQL}
\title{MODEL: mod_sex_null_diff_PQL}
\usage{
data(mod_sex_null_diff_PQL)
}
\description{
Model when the females is focal in intersex interactions.
the model is null (intercept and type of interaction) and is fitted with PQL
}
\keyword{datasets}
