% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_social_null_diff_PQL}
\alias{mod_social_null_diff_PQL}
\title{MODEL: mod_social_null_diff_PQL}
\usage{
data(mod_social_null_diff_PQL)
}
\description{
Model when the individual with the highest social support is focal
in intersex interactions.
the model is null (only intercept and type of interaction) and is fitted with PQL
}
\keyword{datasets}
