% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod_social_null_same_PQL}
\alias{mod_social_null_same_PQL}
\title{MODEL: mod_social_null_diff_PQL
Model when the individul with the highest social support is focal
in intrasex interactions.
the model is null (intercept and type of interaction) and is fitted with PQL}
\usage{
data(mod_social_null_same_PQL)
}
\description{
MODEL: mod_social_null_diff_PQL
Model when the individul with the highest social support is focal
in intrasex interactions.
the model is null (intercept and type of interaction) and is fitted with PQL
}
\keyword{datasets}
