% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{natives_relat}
\alias{natives_relat}
\title{PLOT: natives relat}
\usage{
data(natives_relat)
}
\description{
A dataset containing the relatedness score for natives males at 3
different ages // used for plot1
}
\keyword{datasets}
