% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot1}
\alias{plot1}
\title{plot1}
\usage{
plot1(DF_female, DF_migrant, DF_native)
}
\arguments{
\item{DF_female, }{females_relat}

\item{DF_migrant, }{migrants_relat}

\item{DF_native, }{natives_relat}
}
\description{
Creates the plot1 of the paper.
}
\examples{
data(females_relat)
data(migrants_relat)
data(natives_relat)
plot1(DF_female = females_relat,
     DF_migrant = migrants_relat,
     DF_native = natives_relat)
}
