% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot1_inside}
\alias{plot1_inside}
\title{plot1 panel (used internally)}
\usage{
plot1_inside(DF, legend, labels_x_axis)
}
\arguments{
\item{DF, }{data frame to be plotted: females_relat, migrants_relat or natives_relat}

\item{legend, }{include or not the axis legend (logical)}

\item{labels_x_axis, }{vector of character to specify the axis labels}
}
\description{
Builds the panel of plot1.
}
\examples{
data(females_relat)
plot1_inside(DF = females_relat,
           legend = FALSE,
           labels_x_axis = c("test", "test"))
}
