% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot2}
\alias{plot2}
\title{Plot2}
\usage{
plot2(Mod_so_diff, Mod_so_same, Mod_mass_diff, Mod_mass_same, Mod_sex)
}
\arguments{
\item{Mod_so_diff, }{social model different sex}

\item{Mod_so_same, }{social model same sex}

\item{Mod_mass_diff, }{mass model different sex}

\item{Mod_mass_same, }{mass model same sex}

\item{Mod_sex, }{sex model diff sex}
}
\description{
Creates the second plot of the paper
}
\examples{
data(mod_social_null_diff_PQL)
data(mod_social_null_same_PQL)
data(mod_mass_null_diff_PQL)
data(mod_mass_null_same_PQL)
data(mod_sex_null_diff_PQL)
plot2(
mod_social_null_diff_PQL,
mod_social_null_same_PQL,
mod_mass_null_diff_PQL,
mod_mass_null_same_PQL,
mod_sex_null_diff_PQL
)
}
