% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_intra_inter}
\alias{plot_intra_inter}
\title{Plot prediction for mass and social panels (used internally)}
\usage{
plot_intra_inter(DF)
}
\arguments{
\item{DF}{data frame (created inside the plot2 function)}
}
\description{
Plots the predictions for the mass and social models. These
predictions are computed and formatted inside the plot2 function.
}
