% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_sex}
\alias{plot_sex}
\title{Plot2 sex panel (used internally)}
\usage{
plot_sex(DF)
}
\arguments{
\item{DF, }{a data frame (created inside the plot2 function)}
}
\description{
Plots the predictions for sex model. These
predictions are computed and formated inside the plot2 function
}
