% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{prepare_different_sex}
\alias{prepare_different_sex}
\title{Prepare the data for the interactions between different sexes}
\usage{
prepare_different_sex(DF1)
}
\arguments{
\item{DF1}{The interactions database (d1)}
}
\value{
a filtered data frames
}
\description{
Keeps the interaction between members of different sexes,
selects the variables of interest, renames the type of
interaction and creates a binomial version fo the social support and
delta weight variables.
}
\examples{
data(d1)
data_diff_sex <- prepare_different_sex(d1)
}
