% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_datasets.R
\name{prepare_resid}
\alias{prepare_resid}
\title{Prepare the data for the residency model (SI)}
\usage{
prepare_resid(DF1)
}
\arguments{
\item{DF1}{The interactions database (d1)}
}
\value{
a filtered data frames
}
\description{
Keep the interclan interactions in which the individuals whith the highest
social support is a migrant and its opponent a native.
It also selects the variables of interest, renames the type of
interaction and creates a binomial version fo the social support and
delta weight variables
}
\examples{
data(d1)
data_resid <- prepare_resid(d1)
}
