% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resid_social}
\alias{resid_social}
\title{resid social}
\format{
A data frame with 153 interactions and 10 variables
}
\usage{
data(resid_social)
}
\description{
A dataset containing the interclan interactions in which the migrants have
the highest social support
}
\keyword{datasets}
