% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same_sex_social}
\alias{same_sex_social}
\title{same sex social}
\format{A data frame with 3024 interactions and 10 variables}
\usage{
data(same_sex_social)
}
\description{
A dataset containing the interaction of all same sex interaction with
the individuals with the highest social support as focal
}
\keyword{datasets}
