% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{same_sex_weight}
\alias{same_sex_weight}
\title{same sex weight}
\format{
A data frame with 3024 interactions and 10 variables
}
\usage{
data(same_sex_weight)
}
\description{
A dataset containing the interaction of all same sex interaction with
the individuals with highest body mass as focal
}
\keyword{datasets}
