% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{get_logLik}
\alias{get_logLik}
\title{loglikelyhood}
\usage{
get_logLik(model, name)
}
\arguments{
\item{model}{model to be used}

\item{name}{name of the model}
}
\description{
wrapper arround logLik.
NOTE: Return a tibble so be aware of the automatic rounding
}
\examples{
data(mod_sex_null_diff_PQL)
get_logLik(mod_sex_null_diff_PQL, "sex_null_PQL")
}
