% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot2}
\alias{plot2}
\title{second plot}
\usage{
plot2(DF_female, DF_migrant, DF_native, PDF)
}
\arguments{
\item{DF_female, }{females_relat}

\item{DF_migrant, }{migrants_relat}

\item{DF_native, }{natives_relat}

\item{PDF, }{a logical to return a PDF version of the plot}
}
\description{
Creates the second plot of the paper.
}
\examples{
data(females_relat)
data(migrants_relat)
data(natives_relat)
plot2(DF_female = females_relat,
     DF_migrant = migrants_relat,
     DF_native = natives_relat, 
     PDF = TRUE)
}
